move scripts to top level and address review
