+++ dev-utils/test-utils.js
@@ -28,8 +28,8 @@ const fs = require('fs')
 const webpack = require('webpack')
 const { default: Launcher } = require('@wdio/cli')
 const sauceConnectLauncher = require('sauce-connect-launcher')
+const JasmineRunner = require('jasmine')
 const { singleRunKarma } = require('./karma')
-const { getSauceConnectOptions } = require('./test-config')
 
 function walkSync(dir, filter, filelist) {
   var files = fs.readdirSync(dir)
@@ -161,8 +161,7 @@ function startSelenium(callback, manualStop) {
   )
 }
 
-function runSauceConnect(callback) {
-  const config = getSauceConnectOptions()
+function runSauceConnect(config, callback) {
   return sauceConnectLauncher(config, (err, sauceConnectProcess) => {
     if (err) {
       console.error('Sauce connect Error', err)
@@ -207,10 +206,34 @@ function runE2eTests(configFilePath, runSelenium) {
   }
 }
 
+function runJasmine(specDir, cb) {
+  const jrunner = new JasmineRunner()
+  /**
+   * spec dir is a relative directory path from the current working dir
+   */
+  jrunner.loadConfig({
+    spec_dir: specDir,
+    spec_files: ['*.spec.js'],
+    helpers: ['*.helper.js']
+  })
+  jrunner.onComplete(passed => {
+    if (!passed) {
+      const err = new Error('Jasmine tests failed')
+      // The stack is not useful in this context.
+      err.stack = ''
+      cb(err)
+    } else {
+      cb()
+    }
+  })
+  jrunner.execute()
+}
+
 module.exports = {
   buildE2eBundles,
   runE2eTests,
   runKarma,
+  runJasmine,
   runSauceConnect,
   startSelenium,
   dirWalkSync: walkSync
