+++ packages/rum/package.json
@@ -21,14 +21,13 @@
     "bundlesize": "npm run build && bundlesize",
     "karma": "karma start",
     "karma:coverage": "karma start --coverage",
-    "script": "node ./run-script.js",
-    "serve": "npm run script serveE2e ./ 8000",
+    "script": "node ../../dev-utils/run-script.js",
     "test:node": "npm run script runNodeTests",
     "test:integration": "npm run script runIntegrationTests",
-    "test:unit": "node ../../dev-utils/run-script.js runUnitTests packages/rum",
-    "test:e2e:supported": "node ../../dev-utils/run-script.js runE2eTests packages/rum/wdio.conf.js",
-    "test:e2e:failsafe": "node ../../dev-utils/run-script.js runE2eTests packages/rum/wdio-failsafe.conf.js",
-    "test:sauce": "node ../../dev-utils/run-script.js runSauceTests packages/rum true test:unit test:e2e:supported test:e2e:failsafe",
+    "test:unit": "npm run script runUnitTests packages/rum",
+    "test:e2e:supported": "npm run script runE2eTests packages/rum/wdio.conf.js",
+    "test:e2e:failsafe": "npm run script runE2eTests packages/rum/wdio-failsafe.conf.js",
+    "test:sauce": "npm run script runSauceTests packages/rum true test:unit test:e2e:supported test:e2e:failsafe",
     "test": "npm-run-all build build:e2e test:node test:integration test:sauce"
   },
   "files": [
